creating one component for privacy and policy pages together
